backend/requirements.txt

Django==4.2.1
django-cors-headers==3.14.0
django-storages==1.13.2
graphene-django==3.0.2
graphene-file-upload==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-decouple==3.8
django-storages==1.13.2
s3transfer==0.6.1
whitenoise==6.4.0
boto3==1.26.129